Refactor application structure by creating reusable Navbar and Footer components; remove redundant platform page; add About, Contact, Features, Pricing, and Platform pages with consistent layout and styling.
